fix windows build

+++ src/coreclr/jit/ifconversion.cpp
@@ -790,7 +790,7 @@ static IntConstSelectOper MatchIntConstSelectValues(int64_t trueVal, int64_t fal
     if (trueVal == falseVal + 1)
         return {GT_ADD, TYP_LONG, 0};
 
-    if (trueVal == ((int32_t)falseVal) + 1)
+    if (trueVal == int64_t(int32_t(falseVal) + 1))
         return {GT_ADD, TYP_LONG, 0};
 
     if (falseVal == 0)
@@ -802,26 +802,26 @@ static IntConstSelectOper MatchIntConstSelectValues(int64_t trueVal, int64_t fal
 
         bitIndex = BitOperations::Log2((uint32_t)trueVal);
         assert(bitIndex > 0);
-        if (trueVal == (1 << bitIndex))
+        if (trueVal == int64_t(1 << bitIndex))
             return {GT_LSH, TYP_INT, bitIndex};
     }
 
     if (trueVal == falseVal << 1)
         return {GT_LSH, TYP_LONG, 0};
 
-    if (trueVal == ((int32_t)falseVal) << 1)
+    if (trueVal == int64_t((int32_t(falseVal)) << 1))
         return {GT_LSH, TYP_INT, 0};
 
     if (trueVal == falseVal >> 1)
         return {GT_RSH, TYP_LONG, 0};
 
-    if (trueVal == ((int32_t)falseVal) >> 1)
+    if (trueVal == int64_t(int32_t(falseVal) >> 1))
         return {GT_RSH, TYP_INT, 0};
 
-    if (trueVal == (uint64_t)falseVal >> 1)
+    if (trueVal == int64_t(uint64_t(falseVal) >> 1))
         return {GT_RSZ, TYP_LONG, 0};
 
-    if (trueVal == ((uint32_t)falseVal) >> 1)
+    if (trueVal == int64_t(uint32_t(falseVal) >> 1))
         return {GT_RSZ, TYP_INT, 0};
 
     return {GT_NONE};
